pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-bias-psychologist"
version = "0.1.0"
description = "A diagnostic tool that systematically detects, measures, and tracks cognitive biases in Large Language Models"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "AI Bias Psychologist Team", email = "team@ai-bias-psych.org"},
]
keywords = ["ai", "bias", "psychology", "llm", "cognitive-bias", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # CLI Framework
    "typer[all]>=0.9.0",
    
    # Data Validation & Serialization
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Data Analysis & Statistics
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    
    # LLM API Clients
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "httpx>=0.25.0",
    
    # Database & Storage
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    
    # Configuration & Environment
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    
    # Logging & Monitoring
    "structlog>=23.2.0",
    "rich>=13.7.0",
    
    # Text Processing & NLP
    "nltk>=3.8.1",
    "spacy>=3.7.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    
    # Sentiment Analysis
    "vaderSentiment>=3.3.2",
    "textblob>=0.17.1",
    
    # Async & Concurrency
    "asyncio-throttle>=1.0.2",
    "aiofiles>=23.2.1",
    
    # Utilities
    "click>=8.1.7",
    "tqdm>=4.66.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    
    # Development & Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "jupyter>=1.0.0",
    "ipython>=8.17.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
]

local = [
    "ollama>=0.1.0",
    "psutil>=5.9.0",
]

[project.urls]
Homepage = "https://github.com/ai-bias-psych/ai-bias-psychologist"
Documentation = "https://ai-bias-psychologist.readthedocs.io"
Repository = "https://github.com/ai-bias-psych/ai-bias-psychologist.git"
Issues = "https://github.com/ai-bias-psych/ai-bias-psychologist/issues"

[project.scripts]
ai-bias-psych = "ai_bias_psych.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["ai_bias_psych"]
known_third_party = ["fastapi", "typer", "pydantic", "pandas", "numpy", "scipy"]

# MyPy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "nltk.*",
    "spacy.*",
    "transformers.*",
    "torch.*",
    "vaderSentiment.*",
    "textblob.*",
    "ollama.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "llm: marks tests that require LLM API access",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# FastAPI configuration
[tool.ai_bias_psych]
# Default configuration values
default_model = "gpt-4"
default_temperature = 0.7
default_max_tokens = 1000
log_level = "INFO"
database_url = "sqlite:///./data/database/bias_psych.db"
response_log_dir = "./data/responses"
export_dir = "./data/exports"
